Moves uploads event handlers from auth/events to uploads/events
