+++ assets/scripts/auth/events.js
@@ -4,14 +4,14 @@ const getFormFields = require(`../../../lib/get-form-fields`)
 
 const api = require('./api')
 const ui = require('./ui')
-const store = require('../store')
+const uploadsEvents = require('../uploads/events')
 
 const onSignIn = function (event) {
   event.preventDefault()
   const data = getFormFields(this)
   api.signIn(data)
     .then(ui.signInSuccess)
-    .then(onViewFiles)
+    .then(uploadsEvents.onViewFiles)
     .catch(ui.signInFailure)
 }
 
@@ -47,68 +47,6 @@ const onSignOut = function (event) {
     .catch(ui.signOutFailure)
 }
 
-const onUploadFile = function (event) {
-  event.preventDefault()
-  const data = new FormData(event.target)
-  api.uploadFile(data)
-    .then(ui.uploadFileSuccess)
-    .then(onViewFiles)
-    .catch(ui.uploadFileFailure)
-}
-
-const onViewFiles = function () {
-  return api.viewFiles()
-    .then(ui.viewFilesSuccess)
-    .then(function () {
-      $('.clickable-row').on('click', onRowClick)
-      $('.delete-file-btn').on('click', onDeleteFile)
-    })
-    .catch(ui.viewFilesFailure)
-}
-
-const onDeleteFile = function (event) {
-  event.stopPropagation()
-  const uploadId = event.target.getAttribute('data-id')
-  store.uploadId = uploadId
-  ui.showDeleteModal()
-}
-
-const onDeleteFileConfirm = function () {
-  api.deleteFile(store.uploadId)
-    .then(ui.deleteFileSuccess)
-    .then(onViewFiles)
-    .catch(ui.deleteFileFailure)
-}
-
-const onRowClick = function (event) {
-  onViewFile(event.target.parentNode.id)
-}
-
-const onViewFile = function (id) {
-  api.viewFile(id)
-    .then(ui.viewFileSuccess)
-    .catch(ui.viewFileFailure)
-}
-
-const onViewDelete = function () {
-  event.stopPropagation()
-  ui.showDeleteModal()
-}
-
-const onUpdateFile = function (event) {
-  event.preventDefault()
-  const data = getFormFields(event.target)
-  api.updateFile(data)
-    .then(ui.updateFileSuccess)
-    .then(onViewFiles)
-    .catch(ui.updateFileFailure)
-}
-
-const onFileLookup = function (event) {
-  event.preventDefault()
-  onViewFile($('#file-lookup-id').val())
-}
-
 const addHandlers = function () {
   $('#sign-in').on('submit', onSignIn)
   $('#new-user').on('click', ui.newUser)
@@ -121,13 +59,6 @@ const addHandlers = function () {
   $('#change-password-cancel').on('click', onChangePasswordCancel)
   $('#change-password').on('submit', onChangePassword)
   $('#sign-out').on('click', onSignOut)
-  $('#upload-form').on('submit', onUploadFile)
-  $('#deleteUploadConfirm').on('click', onDeleteFileConfirm)
-  $('#view-delete-button').on('click', onViewDelete)
-  $('#view-back-button').on('click', ui.showHomePage)
-  $('#view-file').on('submit', onUpdateFile)
-  $('#tags').tagsInput()
-  $('#file-lookup').on('click', onFileLookup)
 }
 
 module.exports = {
