+++ assets/scripts/shared/api.js
@@ -1,0 +1,8 @@
+'use strict'
+
+const config = require('../config.js')
+const store = require('../store.js')
+
+module.exports = {
+
+}
