+++ assets/scripts/shared/ui.js
@@ -1,0 +1,12 @@
+'use strict'
+
+const store = require('../store.js')
+
+const clearForm = function (id) {
+  $(id)[0].reset()
+}
+
+
+module.exports = {
+
+}
