+++ assets/scripts/uploads/api.js
@@ -1,0 +1,9 @@
+'use strict'
+
+const config = require('../config.js')
+const store = require('../store.js')
+
+
+module.exports = {
+
+}
