+++ assets/scripts/uploads/ui.js
@@ -1,0 +1,13 @@
+'use strict'
+
+const store = require('../store.js')
+const uploadsTableHandlebar = require('../templates/uploadsTable.handlebars')
+
+const clearForm = function (id) {
+  $(id)[0].reset()
+}
+
+
+module.exports = {
+
+}
